feat: SupplementCardコンポーネント抽出とUI改善完了 - 1785行index.tsxからの分離、容量表示改善、サプリ名オーバーレイ、テキスト省略処理、高さ統一、中央揃え、テスト実装(7ケース)、タイミング/回数ベース対応、アニメーション機能、レスポンシブ対応を含む完全なコンポーネント化
