+++ docs/functional-test-checklist.md
@@ -1,0 +1,140 @@
+# サプリメント管理アプリ 機能テストチェックリスト
+
+## 1. 認証機能
+
+- [ ] **1.1** ログインページで正しい認証情報を入力してログインできる
+- [ ] **1.2** 新規登録ページでアカウントを作成できる
+- [ ] **1.3** ログアウト機能が正常に動作する
+- [ ] **1.4** 未認証時は認証ページにリダイレクトされる
+
+## 2. サプリメント追加・編集機能
+
+### 2.1 基本情報入力
+
+- [ ] **2.1.1** サプリメント名を入力できる
+- [ ] **2.1.2** 内容量と単位を入力できる（錠、粒、ml、gなど）
+- [ ] **2.1.3** 1回の服用量と単位を入力できる
+- [ ] **2.1.4** 画像をアップロードできる
+
+### 2.2 服用管理方法設定
+
+- [ ] **2.2.1** 「タイミングベース」と「回数ベース」から選択できる
+- [ ] **2.2.2** タイミングベース選択時、朝・昼・夜から少なくとも1つ選択必須
+- [ ] **2.2.3** 回数ベース選択時、1日の目標服用回数を入力必須
+
+### 2.3 推奨服用方法設定
+
+- [ ] **2.3.1** 食前・食後・空腹時・なしから1つ選択できる（排他的選択）
+- [ ] **2.3.2** 就寝前はチェックボックスで独立して選択できる
+- [ ] **2.3.3** 詳細情報（情報ボタン）を押すと説明が表示される
+
+## 3. カードUI表示機能
+
+### 3.1 基本情報表示
+
+- [ ] **3.1.1** サプリメント名が表示される
+- [ ] **3.1.2** 残り容量が正確に表示される
+- [ ] **3.1.3** 1回の服用量が表示される
+- [ ] **3.1.4** 画像が表示される（ない場合は"no-image"）
+
+### 3.2 推奨服用方法表示
+
+- [ ] **3.2.1** 推奨服用方法が設定されている場合、「服用方法」セクションが表示される
+- [ ] **3.2.2** 設定された推奨服用方法（食前・食後・空腹時・就寝前）のアイコンとラベルが表示される
+- [ ] **3.2.3** 推奨服用方法が設定されていない場合、「服用方法」セクションは表示されない
+
+### 3.3 タイミングベース表示
+
+- [ ] **3.3.1** タイミングベース選択時、設定されたタイミング（朝・昼・夜）のボタンのみ表示される
+- [ ] **3.3.2** 未設定のタイミングボタンは表示されない
+- [ ] **3.3.3** タイミングボタンの状態（未服用・服用済み）が正確に表示される
+
+### 3.4 回数ベース表示
+
+- [ ] **3.4.1** 回数ベース選択時、プラス・マイナスボタンが表示される
+- [ ] **3.4.2** 現在の服用回数が数字で表示される
+- [ ] **3.4.3** 目標服用回数が表示される（例：「目標: 2 / 3 回」）
+- [ ] **3.4.4** 服用履歴が横スクロールで表示される
+
+## 4. 服用記録機能
+
+### 4.1 タイミングベース服用記録
+
+- [ ] **4.1.1** タイミングボタンをクリックすると服用記録される
+- [ ] **4.1.2** 服用記録時、残り容量が減る（残り容量 = 残り容量 - 1回服用量）
+- [ ] **4.1.3** 既に服用済みのタイミングをクリックすると服用記録が取り消される
+- [ ] **4.1.4** 服用取り消し時、残り容量が戻る（残り容量 = 残り容量 + 1回服用量）
+- [ ] **4.1.5** 残り容量が不足している場合、服用ボタンが無効化される
+
+### 4.2 回数ベース服用記録
+
+- [ ] **4.2.1** プラスボタンをクリックすると服用回数が増加する
+- [ ] **4.2.2** 服用回数増加時、残り容量が減る（残り容量 = 残り容量 - 1回服用量）
+- [ ] **4.2.3** マイナスボタンをクリックすると服用回数が減少する
+- [ ] **4.2.4** 服用回数減少時、残り容量が戻る（残り容量 = 残り容量 + 1回服用量）
+- [ ] **4.2.5** 残り容量が不足している場合、プラスボタンが無効化される
+- [ ] **4.2.6** 服用回数が0の場合、マイナスボタンが無効化される
+
+### 4.3 アニメーションフィードバック
+
+- [ ] **4.3.1** 服用記録時にアニメーションフィードバックが表示される
+- [ ] **4.3.2** アニメーション表示中は操作ボタンが無効化される
+- [ ] **4.3.3** アニメーション完了後、服用状態が正確に反映される
+
+## 5. 日付変更時のリセット機能
+
+- [ ] **5.1** 日付が変わると服用状況がリセットされる
+- [ ] **5.2** タイミングベースの場合、すべてのタイミングが未服用状態になる
+- [ ] **5.3** 回数ベースの場合、服用回数が0にリセットされる
+- [ ] **5.4** 残り容量は日付変更時に変更されない
+
+## 6. データ永続化
+
+- [ ] **6.1** 服用記録がFirestoreに正確に保存される
+- [ ] **6.2** ページをリロードしても服用状態が維持される
+- [ ] **6.3** ログアウト・再ログインしても服用状態が維持される
+
+## 7. エラーハンドリング
+
+- [ ] **7.1** ネットワークエラー時に適切なエラーメッセージが表示される
+- [ ] **7.2** バリデーションエラー時に適切なエラーメッセージが表示される
+- [ ] **7.3** 操作失敗時にUIが適切に復元される
+
+## 8. レスポンシブデザイン
+
+- [ ] **8.1** モバイルデバイスで正常に表示・操作できる
+- [ ] **8.2** タブレットデバイスで正常に表示・操作できる
+- [ ] **8.3** デスクトップで正常に表示・操作できる
+
+## 9. アクセシビリティ
+
+- [ ] **9.1** キーボードナビゲーションが正常に動作する
+- [ ] **9.2** ARIA属性が適切に設定されている
+- [ ] **9.3** スクリーンリーダーで内容が読み上げられる
+
+## 10. パフォーマンス
+
+- [ ] **10.1** ページ読み込み時間が適切（3秒以内）
+- [ ] **10.2** 操作に対するレスポンスが適切（1秒以内）
+- [ ] **10.3** 大量のサプリメントがある場合でもスムーズに動作する
+
+---
+
+## 現在の優先修正項目
+
+### 高優先度（すぐに修正が必要）
+
+- [ ] **2.3.1** 推奨服用方法がカードUIに表示されない問題
+- [ ] **3.3.1** タイミングベースで全てのタイミングボタンが表示される問題
+- [ ] **4.1.2** タイミングボタンクリック時に残り容量が反映されない問題
+- [ ] **4.2.2** 回数ベースで残り容量が反映されない問題
+
+### 中優先度（改善が望ましい）
+
+- [ ] **3.2.1** 推奨服用方法の表示スタイル調整
+- [ ] **4.3.1** アニメーションフィードバックの改善
+
+### 低優先度（将来的な改善）
+
+- [ ] **7.1** エラーハンドリングの強化
+- [ ] **10.1** パフォーマンス最適化
